feat(operations): centralize mutation payload builders
Centralize workout, routine, and measurement payload semantics in operations.
