+++ .changeset/curvy-builders-share.md
@@ -1,0 +1,9 @@
+---
+"@hevy-mcp/operations": patch
+"@hevy-mcp/core": patch
+"hevy-mcp": patch
+"@hevy-mcp/worker": patch
+"@chrisdoc/hevy-cli": patch
+---
+
+Move shared mutation payload builders and Hevy mutation quirks into operations.
